Force .nojekyll creation always
GitHub Pages uses Jekyll to pre-process gh-pages repositories to convert
Markdown into HTML pages and so forth, but this tool ignores all
directory entries which begin with an underscore.  Sphinx puts all its
static content in the `_static` directory, so the easiest solution is
just to turn off Jekyll so that GitHub correctly hosts our stylesheets.

This is done simply by dropping the empty file `.nojekyll` into the
repository root, and ensuring that it's copied into the correct place as
part of the Sphinx build process.